no commit message given

+++ plugins/code.sh
@@ -5,8 +5,8 @@ MSKEYID="EB3E94ADBE1229CF"
 MSAPT="/etc/apt/sources.list.d/code.sources"
 
 function _install_code() {
-	gpg --keyserver "keyserver.ubuntu.com" --recv-keys "${MSKEYID}"
-	gpg --export --armor "${MSKEYID}" | sudo gpg --dearmor -o "${MSRING}"
+	sudo gpg --keyserver "keyserver.ubuntu.com" --recv-keys "${MSKEYID}"
+	sudo gpg --export --armor "${MSKEYID}" | sudo gpg --dearmor -o "${MSRING}"
 	sudo tee "${MSAPT}" > /dev/null << EOF
 Types: deb
 URIs: https://packages.microsoft.com/repos/code
@@ -16,12 +16,29 @@ Architectures: amd64
 Signed-By: ${MSRING}
 EOF
 	sudo apt update
-	sudo apt install code
+	sudo DEBIAN_FRONTEND=noninteractive apt-get install -y code
+	# sudo apt install code
 }
 
 function _uninstall_code() {
-	sudo rm "${MSAPT}"
-	sudo rm "${MSRING}"
+	if sudo gpg --list-keys "${MSKEYID}" &> /dev/null; then
+		echo "Key ${MSKEYID} found in personal keyring. Deleting..."
+		sudo gpg --delete-keys --yes "${MSKEYID}"
+	else
+		echo "Key ${MSKEYID} not found in personal keyring. Nothing to do."
+	fi
+	if [ -f "${MSAPT}" ]
+	then
+		sudo rm "${MSAPT}"
+	else
+		echo "file [${MSAPT}] is not there, not removing"
+	fi
+	if [ -f "${MSRING}" ]
+	then
+		sudo rm "${MSRING}"
+	else
+		echo "file [${MSRING}] is not there, not removing"
+	fi
 	sudo dpkg --purge code
 	sudo apt update
 }
